@@ -16,12 +16,13 @@
   "author": "leader22",
   "license": "MIT",
   "dependencies": {
-    "chart.js": "^2.1.2",
+    "chart.js": "^2.6.0",
     "object-assign": "^4.1.0",
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2",
-    "react-router": "^2.4.0",
-    "react-tap-event-plugin": "^1.0.0"
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-router-dom": "^4.1.1",
+    "react-tap-event-plugin": "^2.0.1"
   },
   "devDependencies": {
     "babel-preset-es2015": "^6.6.0",
