@@ -1,4 +1,5 @@
 const React = require('react');
+const PropTypes = require('prop-types');
 
 const KDInput = ({
   kill, death,
@@ -24,9 +25,9 @@ const KDInput = ({
 };
 
 KDInput.propTypes = {
-  kill:     React.PropTypes.string.isRequired,
-  death:    React.PropTypes.string.isRequired,
-  onChange: React.PropTypes.func.isRequired,
+  kill:     PropTypes.string.isRequired,
+  death:    PropTypes.string.isRequired,
+  onChange: PropTypes.func.isRequired,
 };
 
 module.exports = KDInput;
