@@ -1,4 +1,5 @@
 const React = require('react');
+const PropTypes = require('prop-types');
 
 let _timer = null;
 
@@ -63,9 +64,9 @@ class SaveBtn extends React.Component {
 }
 
 SaveBtn.propTypes = {
-  canInput:    React.PropTypes.bool.isRequired,
-  onSave:      React.PropTypes.func.isRequired,
-  onAfterSave: React.PropTypes.func,
+  canInput:    PropTypes.bool.isRequired,
+  onSave:      PropTypes.func.isRequired,
+  onAfterSave: PropTypes.func,
 };
 
 SaveBtn.defaultProps = {
